Revert "Reshaping e-venv tests and adding extension install test"
This reverts commit bf14a274e69ddd868e9dbf23a2b971c91cb9a564.
