+++ Jenkinsfile
@@ -38,7 +38,8 @@ pipeline {
                     npm i --also=dev
                     rm -rf $INMANTA_EXTENSION_TEST_ENV
                     python3 -m venv $INMANTA_EXTENSION_TEST_ENV
-                    INMANTA_PYTHON_PATH="$INMANTA_EXTENSION_TEST_ENV/bin/python3" INMANTA_COMPILER_VENV="$WORKSPACE/compiler-venv" xvfb-run npm run test
+                    $INMANTA_EXTENSION_TEST_ENV/bin/python3 -m pip install -e $INMANTA_LS_PATH
+                    INMANTA_PYTHON_PATH="$INMANTA_EXTENSION_TEST_ENV/bin/python3" xvfb-run npm run test
                 '''
             }
         }
